% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMLi.R
\name{EMMLi}
\alias{EMMLi}
\title{Evaluating modularity with maximum likelihood}
\usage{
EMMLi(corr, N_sample, mod, saveAs = NULL, abs = TRUE)
}
\arguments{
\item{corr}{Lower triangle or full correlation matrix. n x n square matrix for n landmarks.}

\item{N_sample}{The number of samples used}

\item{mod}{A data frame defining the models. The first column should contain the landmark names. Subsequent columns should define which landmarks are contained with which module. If a landmark should be ignored for a specific model, the element should be NA.}

\item{saveAs}{A character string defining the filename and path for where to save output. If NULL, the output is not saved to file.}

\item{abs}{Logical denoting whether absolute values should be used. TODO anjali}
}
\value{
A list containing two elements. The first (results) gives the AIC results for each model.
 The second (rho) gives the within and between module correlations.
 Optionally, the output is saved to the file defined by the saveAs argument with only models with a 
 posterior probability > 0.01 being saved.
}
\description{
Calculates the AICc values of different models of modularity.
}
\details{
The publication describing this analysis is A. Goswami1 and J. Finarelli
   (2016) EMMLi: A maximum likelihood approach to the analysis of modularity.
   Evolution \url{http://onlinelibrary.wiley.com/doi/10.1111/evo.12956/abstract}.
}
\examples{
 set.seed(1)

 # Chose a filename and directory for output
 file <- paste0(tempdir(), 'EMMLiTest.csv')
 
 # Create a correlation matrix
 dat <- matrix(runif(36, -1, 1), ncol = 6, nrow = 6)
 diag(dat) <- 1

 # Create dataframe of model definitions.
 #   Use NAs indicate landmarks which are not part of the model
 mod1 <- data.frame(landmarks = letters[1:6], 
              modelA = rep(c(1, 2), each = 3),
              modelB = rep(c(1,2), 3),
              model3= c(rep(c(1,2), 2), NA, NA)) 


 # run EMMLi
 EMMLi(dat, 20, mod1, file)

 unlink(file)

 # run EMMLi without writing output
 EMMLi(dat, 20, mod1)
}

