% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMLipackage.R
\docType{package}
\name{EMMLipackage}
\alias{EMMLipackage}
\alias{EMMLipackage-package}
\title{EMMLi: A Maximum Likelihood Approach To The Analysis Of Modularity}
\description{
The only function is \code{\link{EMMLi}}. This function takes a covariance
 matrix and a data frame that defines the which module each landmark is a 
 a part of for each model.
}

