% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macacaCorrel}
\alias{macacaCorrel}
\title{Correlation matrix for 61 landmarks from Japanese macaque 
  (\emph{Macaca fuscata}) craniums.}
\format{A 61 x 61 matrix}
\source{
Goswami, A. and P. D. Polly. 2010. Methods for studying morphological integration and
   modularity. Pp. 213-243 in J. Alroy, and E. G. Hunt, eds. Quantitative Methods in
   Paleobiology. Paleontological Society Special Publications.
}
\usage{
macacaCorrel
}
\description{
3D coordinates for 61 landmarks on the crania of 42 juvenile Japanese macaque 
  (\emph{Macaca fuscata}) from the Primate Research Institute at Inuyama, 
  Japan, was aligned and scaled with Generalized Procrustes superimposition. 
  The vector congruence coefficient correlation was then calculated for each pair
  of landmarks
}
\details{
#'@seealso \code{\link{macacaModels}}
}
\keyword{datasets}

