% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMLipackage.R
\docType{package}
\name{EMMLipackage}
\alias{EMMLipackage}
\alias{EMMLipackage-package}
\title{EMMLi: A Maximum Likelihood Approach To The Analysis Of Modularity}
\description{
The only function is \code{\link{EMMLi}}. This function takes a covariance
 matrix and a data frame that defines the which module each landmark is a 
 a part of for each model.
}
\details{
The publication describing this analysis is A. Goswami1 and J. Finarelli
   (2016) EMMLi: A maximum likelihood approach to the analysis of modularity.
   Evolution \url{http://onlinelibrary.wiley.com/doi/10.1111/evo.12956/abstract}.
}

